+++ src/python/k4a/src/k4a/_bindings/k4a.py
@@ -182,7 +182,7 @@ k4a_image_create_from_buffer = _k4a_lib.k4a_image_create_from_buffer
 k4a_image_create_from_buffer.restype = EStatus
 k4a_image_create_from_buffer.argtypes=(
     _ctypes.c_int, _ctypes.c_int, _ctypes.c_int, _ctypes.c_int, _ctypes.POINTER(_ctypes.c_uint8),
-    _ctypes.c_ulonglong, _memory_allocate_cb, _ctypes.c_void_p, _ctypes.POINTER(_ImageHandle))
+    _ctypes.c_size_t, _memory_allocate_cb, _ctypes.c_void_p, _ctypes.POINTER(_ImageHandle))
 
 
 
@@ -194,7 +194,7 @@ k4a_image_get_buffer.argtypes=(_ImageHandle,)
 
 #K4A_EXPORT size_t k4a_image_get_size(k4a_image_t image_handle);
 k4a_image_get_size = _k4a_lib.k4a_image_get_size
-k4a_image_get_size.restype = _ctypes.c_ulonglong
+k4a_image_get_size.restype = _ctypes.c_size_t
 k4a_image_get_size.argtypes=(_ImageHandle,)
 
 
@@ -324,7 +324,7 @@ k4a_device_stop_imu.argtypes = (_DeviceHandle,)
 k4a_device_get_serialnum = _k4a_lib.k4a_device_get_serialnum
 k4a_device_get_serialnum.restype = EBufferStatus
 k4a_device_get_serialnum.argtypes = (_DeviceHandle, 
-    _ctypes.POINTER(_ctypes.c_char), _ctypes.POINTER(_ctypes.c_ulonglong))
+    _ctypes.POINTER(_ctypes.c_char), _ctypes.POINTER(_ctypes.c_size_t))
 
 
 #K4A_EXPORT k4a_status_t k4a_device_get_version(k4a_device_t device_handle, HardwareVersion *version);
@@ -373,7 +373,7 @@ k4a_device_set_color_control.argtypes = (_DeviceHandle, _ctypes.c_int, _ctypes.c
 #                                                              size_t *data_size);
 k4a_device_get_raw_calibration = _k4a_lib.k4a_device_get_raw_calibration
 k4a_device_get_raw_calibration.restype = EBufferStatus
-k4a_device_get_raw_calibration.argtypes = (_DeviceHandle, _ctypes.POINTER(_ctypes.c_uint8), _ctypes.POINTER(_ctypes.c_ulonglong))
+k4a_device_get_raw_calibration.argtypes = (_DeviceHandle, _ctypes.POINTER(_ctypes.c_uint8), _ctypes.POINTER(_ctypes.c_size_t))
 
 
 #K4A_EXPORT k4a_status_t k4a_device_get_calibration(k4a_device_t device_handle,
@@ -401,7 +401,7 @@ k4a_device_get_sync_jack.argtypes = (_DeviceHandle, _ctypes.POINTER(_ctypes.c_bo
 k4a_calibration_get_from_raw = _k4a_lib.k4a_calibration_get_from_raw
 k4a_calibration_get_from_raw.restype = EStatus
 k4a_calibration_get_from_raw.argtypes = (_ctypes.POINTER(_ctypes.c_char), 
-    _ctypes.c_ulonglong, _ctypes.c_int, _ctypes.c_int, _ctypes.POINTER(_Calibration))
+    _ctypes.c_size_t, _ctypes.c_int, _ctypes.c_int, _ctypes.POINTER(_Calibration))
 
 
 #K4A_EXPORT k4a_status_t k4a_calibration_3d_to_3d(const k4a_calibration_t *calibration,
